notifications: Split out message_notifications module.
