+++ tools/test-js-with-node
@@ -133,6 +133,7 @@ EXEMPT_FILES = make_set(
         "web/src/message_list_view.js",
         "web/src/message_lists.js",
         "web/src/message_live_update.js",
+        "web/src/message_notifications.js",
         "web/src/message_scroll.js",
         "web/src/message_scroll_state.ts",
         "web/src/message_util.js",
